feat: módulo de inserção para insertion sort
Criação de módulo que realiza a inserção e deslocamento que será
utilizado como auxiliar para o algoritmo de insertion sort.
 Changes to be committed:
	new file:   argmin/README.md
	modified:   build.ps1
	modified:   build.sh
	new file:   gtkw/insere.gtkw
	new file:   insere/README.md
	new file:   insere/insere.vhdl
	new file:   insere/insere_tb.vhdl
	modified:   norm/README.md
	modified:   sqrt/README.md

+++ build.sh
@@ -58,15 +58,19 @@ for MODULO in ${MODULOS[@]}; do
     ghdl -e ${GHDLFLAGS} ${MODULO}_tb
 
     # Executa bancada de testes
-    echo -e "${VERMELHO}  ghdl -r ${GHDLFLAGS} ${MODULO}_tb --vcd=${MODULO}_tb.vcd"
-    ghdl -r ${GHDLFLAGS} ${MODULO}_tb --vcd=${MODULO}_tb.vcd > /dev/null
+    echo -e "${VERMELHO}  ghdl -r ${GHDLFLAGS} ${MODULO}_tb --wave=${MODULO}_tb.ghw"
+    ghdl -r ${GHDLFLAGS} ${MODULO}_tb --wave=${MODULO}_tb.ghw > /dev/null
 
     mkdir -p ../resultados/${MODULO}
 
-    echo -e "${CIANO}  mv ${MODULO}.csv ../resultados/${MODULO}/${MODULO}.csv"
-    mv ${MODULO}.csv ../resultados/${MODULO}/${MODULO}.csv
-    echo -e "  mv ${MODULO}.vcd ../resultados/${MODULO}/${MODULO}.vcd"
-    mv ${MODULO}_tb.vcd ../resultados/${MODULO}/${MODULO}_tb.vcd
+    if test -f "${MODULO}.csv"; then
+        echo -e "${CIANO}  mv ${MODULO}.csv ../resultados/${MODULO}/${MODULO}.csv"
+        mv ${MODULO}.csv ../resultados/${MODULO}/${MODULO}.csv
+    fi;
+    if test -f "${MODULO}_tb.ghw"; then
+        echo -e "  mv ${MODULO}.ghw ../resultados/${MODULO}/${MODULO}_tb.ghw"
+        mv ${MODULO}_tb.ghw ../resultados/${MODULO}/${MODULO}_tb.ghw
+    fi;
 
 done
